+++ spec/graphql/relay/global_node_identification_spec.rb
